requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
pymongo==4.6.0
passlib==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0
starlette==0.27.0
google-generativeai==0.8.6
PyPDF2==3.0.1
Pillow==10.1.0
pytesseract==0.3.10
python-docx==1.1.0
python-pptx==0.6.21
groq==0.3.0
fastapi-mail==1.4.2
